update 1
Create the Basic mechanic of the car movement and jump method using the wheel collider, which is built in unity.
create the car physics component - the rigid body component that gives the object a physical movement around the environment that is affected by gravity.
create the camera to follow a script to follow that follows the car in the x, and y-axis offset from the car and stays on the same z-axis.
create a basic terrain and make a layer called "Ground" to make the car know to recognize where is the ground surface to avoid the car to unlimited jump in the air by using Raycast