update 2
limit the car rotation in 45 and -45 in y axis rotation euler angles to make the car to keep moving forward across the road.
auto rotate the car toward to direction of the road if not pressing nothing or rich the boundries on the sides.
add points lights to the car,points lights and directional light to the environment for better visual.
using particle system component to emit particles from the exhust in a relations to the speed of the car.
using Trail Renderer component to see where the car is crossing.
add boundries by clamp the z rotation of the car to avoid the player get out from the road