update 4
i create a gun shooting mechanic script, using Transform.translate to move the gun in the desired direction. The shooting action should be triggered by pressing the "F" key. Instantiate a bullet every 0.5 seconds. To visualize the bullet's trajectory, add a Trail Renderer component to it. When the car enters the Dead zone or colide with the walls activate the onTriggerEnter event to destroy the car.