update 5
First, I set the mechanic of the camera to top down perspective .
Also , I Created a shooter that fires balls from the air by using the Rigidbody component. The shooter detects the player using Physics.OverlapSphere and the layer associated with the player. It aims directly at the player with a random ratio between the minimum and maximum values.
In Addition , Create a shield box that activates a shield on the car when the car collects the box using the OnTriggerEnter event to detect the player. When the shield is activated, the boolean value inside the car script changes. The shield remains active for a random duration defined within the shieldBox script, and it is deactivated using the Invoke built-in method in Unity. Set the shield within the obstacles script to protect the player from any obstacles in the scene.